+++ core/src/main/kotlin/com/gitlab/kordlib/core/entity/PermissionOverwriteEntity.kt
@@ -57,7 +57,7 @@ class PermissionOverwriteEntity(
      *
      * @throws [RequestException] if anything went wrong during the request.
      */
-    suspend fun getGuildOrNull(): Guild? = supplier.getGuild(guildId)
+    suspend fun getGuildOrNull(): Guild? = supplier.getGuildOrNull(guildId)
 
     /**
      * Requests to delete this overwrite.