+++ core/src/main/kotlin/com/gitlab/kordlib/core/supplier/CacheEntitySupplier.kt
@@ -238,4 +238,6 @@ class CacheEntitySupplier(private val kord: Kord) : EntitySupplier {
 
         return User(data, kord)
     }
+
+
 }